Calling of new endpoint moved out to own file
